Rollout v0 and wait for next release for bugfix Bugs 1. The message from client A sometimes is received by client B, sometimes not.
