+++ README.md
@@ -1,3 +1,6 @@
 # grpc-chat-app
-![](grpc.png)
 gRPC based terminal-2-terminal based chat app
+
+![](grpc.png)
+
+
